==> notebooks/14_inference_pipeline.ipynb <==
%load_ext autoreload
%autoreload 2

import hopsworks
import src.config as config
from src.feature_store_api import get_feature_store, get_or_create_feature_view
from src.config import FEATURE_VIEW_METADATA
from src.data import add_missing_slots
from src.inference import load_batch_of_features_from_store
from datetime import datetime, timedelta
import pandas as pd
from src.config import FEATURE_GROUP_NAME
from src.config import FEATURE_GROUP_VERSION

# # connect to the project
# project = hopsworks.login(
#     project=config.HOPSWORKS_PROJECT_NAME,
#     api_key_value=config.HOPSWORKS_API_KEY
# )

# # connect to the feature store
# feature_store = project.get_feature_store()

# # connect to the feature group
# feature_group = feature_store.get_feature_group(
#     name=FEATURE_GROUP_NAME,
#     version=FEATURE_GROUP_VERSION
# )

# # read the feature group data
# feature_group_data = feature_group.read()

# # print the data
# print(feature_group_data)


# feature_group_data.dtypes

# feature_group_data

# # Convert current_date, fetch_data_from, and fetch_data_to to UTC and floor to the nearest hour
# current_date = pd.to_datetime(datetime.utcnow(), utc=True).floor('H')
# fetch_data_from = current_date - pd.Timedelta(days=28)
# fetch_data_to = current_date - pd.Timedelta(hours=1)

# # Filter feature_group_data based on pickup_hour
# if feature_group_data.pickup_hour.dt.tz is None:
#     feature_group_data.pickup_hour = feature_group_data.pickup_hour.dt.tz_localize('UTC')
# filtered_data = feature_group_data[feature_group_data.pickup_hour.between(fetch_data_from,fetch_data_to)]

# print(filtered_data.pickup_hour.min())
# print(filtered_data.pickup_hour.max())


import src.config as config
from src.feature_store_api import get_feature_store, get_or_create_feature_view
from src.config import FEATURE_VIEW_METADATA
from src.data import add_missing_slots
from src.inference import load_batch_of_features_from_store
from datetime import datetime, timedelta
import pandas as pd


current_date = pd.to_datetime(datetime.utcnow(), utc=True).floor('H')
print(f'{current_date=}')

# feature_view = get_or_create_feature_view(FEATURE_VIEW_METADATA)

# # fetch data from the feature store
# fetch_data_from = current_date - timedelta(days=28)
# fetch_data_to = current_date - timedelta(hours=1)

# # add plus minus margin to make sure we do not drop any observation
# ts_data = feature_view.get_batch_data(
#     start_time=fetch_data_from - timedelta(days=1),
#     end_time=fetch_data_to + timedelta(days=1)
# )

# if ts_data.pickup_hour.dt.tz is None:
#     ts_data.pickup_hour = ts_data.pickup_hour.dt.tz_localize('UTC')
# ts_data = ts_data[ts_data.pickup_hour.between(fetch_data_from,fetch_data_to)]

# # ts_data = add_missing_slots(ts_data)

# # filter data to the time period we are interested in
# # ts_data = add_missing_slots(ts_data)

# # sort data by location and time
# ts_data.sort_values(by=['pickup_location_id', 'pickup_hour'], inplace=True)

# print(ts_data.pickup_hour.min())
# print(ts_data.pickup_hour.max())
# print(len(ts_data.pickup_hour.unique()))
# print(len(ts_data.pickup_location_id.unique()))

features = load_batch_of_features_from_store(current_date)

from src.inference import (
    load_model_from_registry,
    get_model_predictions
)

model = load_model_from_registry()
predictions = get_model_predictions(model, features)

predictions['pickup_hour'] = current_date
predictions

### Save these predictions in the feature store, so they can be later consumed by our Streamlit app

from src.feature_store_api import get_feature_store
import src.config as config

# connect to the feature group
feature_group = get_feature_store().get_or_create_feature_group(
    name=config.FEATURE_GROUP_MODEL_PREDICTIONS,
    version=1,
    description="Predictions generate by our production model",
    primary_key = ['pickup_location_id', 'pickup_hour'],
    event_time='pickup_hour',
)

feature_group.insert(predictions, write_options={"wait_for_job": True})
